// src/enums/cefrLevels.ts
export enum CefrValue {
  A0 = 0,
  A1 = 10,
  A1_PLUS = 15,
  A2 = 25,
  A2_PLUS = 35,
  B1 = 50,
  B1_PLUS = 60,
  B2 = 75,
  B2_PLUS = 85,
  C1 = 95,
  C1_PLUS = 100,
}

export enum CefrLabel {
  A0 = 'A0',
  A1 = 'A1',
  A1_PLUS = 'A1+',
  A2 = 'A2',
  A2_PLUS = 'A2+',
  B1 = 'B1',
  B1_PLUS = 'B1+',
  B2 = 'B2',
  B2_PLUS = 'B2+',
  C1 = 'C1',
  C1_PLUS = 'C1+',
}